+++ CHANGELOG.md
@@ -19,6 +19,7 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 **Other changes:**
 
 * The [mse234680](https://meta.stackexchange.com/q/234680) fix now also fixes IRIs pasted into the edit box. (EXPERIMENTAL!)
+* Rewrote the [mse153528](https://meta.stackexchange.com/q/153528) fix to work with the new "toast" style notification boxes.
 
 
 1.54 (2019-03-08)