+++ CHANGELOG.md
@@ -21,6 +21,7 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 * The [mse234680](https://meta.stackexchange.com/q/234680) fix now also fixes IRIs pasted into the edit box. (EXPERIMENTAL!)
 * Rewrote the [mse153528](https://meta.stackexchange.com/q/153528) fix to work with the new "toast" style notification boxes.
 * The [mse74274](https://meta.stackexchange.com/q/74274) fix stopped working after changes to SE code, made it work again.
+* Tweak the [mse172931](https://meta.stackexchange.com/q/172931) fix to correctly hide vote buttons on answers in review again.
 
 
 1.54 (2019-03-08)