+++ CHANGELOG.md
@@ -21,8 +21,9 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 * The [mse234680](https://meta.stackexchange.com/q/234680) fix now also fixes IRIs pasted into the edit box. (EXPERIMENTAL!)
 * Rewrote the [mse153528](https://meta.stackexchange.com/q/153528) fix to work with the new "toast" style notification boxes.
 * The [mse74274](https://meta.stackexchange.com/q/74274) fix stopped working after changes to SE code, made it work again.
-* Tweak the [mse172931](https://meta.stackexchange.com/q/172931) fix to correctly hide vote buttons on answers in review again.
-* Tweak the [mse291623](https://meta.stackexchange.com/q/291623) CSS selector to only apply to the mobile view. (Avoids color mismatch for italic text in visited links on Chrome.)
+* Tweaked the [mse172931](https://meta.stackexchange.com/q/172931) fix to correctly hide vote buttons on answers in review again.
+* Tweaked the [mse291623](https://meta.stackexchange.com/q/291623) CSS selector to only apply to the mobile view. (Avoids color mismatch for italic text in visited links on Chrome.)
+* Added CSS to stop the [mso287222](https://meta.stackoverflow.com/q/287222) fix from pushing the new activity indicator on custom question filters to a separate line ([issue #44](https://github.com/vyznev/soup/issues/44)).
 
 
 1.54 (2019-03-08)